download_model.py
import requests
import os

def download_asl_model():
    """
    Download a pre-trained ASL model from available sources
    """
    print("Searching for available ASL models...")
    
    # List of potential model sources
    model_urls = [
        # Add actual URLs when found
        "https://github.com/your-repo/asl-model.h5",  # Replace with actual URL
    ]
    
    print("To get a proper trained model:")
    print("1. Search GitHub for 'ASL alphabet CNN model .h5'")
    print("2. Check Kaggle datasets for ASL models")
    print("3. Look for Google Drive links in ASL projects")
    print("4. Train your own model with the ASL dataset")
    
    print("\nDataset sources:")
    print("- Kaggle: ASL Alphabet Dataset")
    print("- GitHub: Sign Language MNIST")
    print("- Roboflow: ASL datasets")
    
    return False

if __name__ == "__main__":
    download_asl_model()