requirements_py37.txt
Django==3.2.25
tensorflow-cpu==2.8.0
opencv-python-headless==4.5.5.62
Pillow==9.0.1
numpy==1.21.6
keras==2.8.0
gunicorn==20.1.0
whitenoise==6.2.0
protobuf==3.20.3